feat(chartcuterie): abstract data normalization
