+++ static/app/components/events/eventStatisticalDetector/lineChart.tsx
@@ -4,7 +4,7 @@ import {useTheme} from '@emotion/react';
 import ChartZoom from 'sentry/components/charts/chartZoom';
 import {LineChart as EChartsLineChart} from 'sentry/components/charts/lineChart';
 import getBreakpointChartOptionsFromData from 'sentry/components/events/eventStatisticalDetector/breakpointChartOptions';
-import type {PageFilters} from 'sentry/types';
+import type {EventsStatsData, PageFilters} from 'sentry/types';
 import type {Series} from 'sentry/types/echarts';
 import useRouter from 'sentry/utils/useRouter';
 import type {NormalizedTrendsTransaction} from 'sentry/views/performance/trends/types';
@@ -12,16 +12,26 @@ import type {NormalizedTrendsTransaction} from 'sentry/views/performance/trends/
 interface ChartProps {
   datetime: PageFilters['datetime'];
   evidenceData: NormalizedTrendsTransaction;
-  percentileSeries: Series[];
+  // TODO @athena: Refactor functionBreakpointChart to use percentileData
+  percentileData?: EventsStatsData;
+  percentileSeries?: Series[];
 }
 
-function LineChart({datetime, percentileSeries, evidenceData}: ChartProps) {
+function LineChart({
+  datetime,
+  percentileData,
+  percentileSeries,
+  evidenceData,
+}: ChartProps) {
   const theme = useTheme();
   const router = useRouter();
 
   const {series, chartOptions} = useMemo(() => {
-    return getBreakpointChartOptionsFromData({percentileSeries, evidenceData}, theme);
-  }, [percentileSeries, evidenceData, theme]);
+    return getBreakpointChartOptionsFromData(
+      {percentileData, percentileSeries, evidenceData},
+      theme
+    );
+  }, [percentileData, percentileSeries, evidenceData, theme]);
 
   return (
     <ChartZoom router={router} {...datetime}>
