+++ static/app/views/performance/transactionSummary/transactionOverview/trendChart/content.tsx
@@ -20,7 +20,7 @@ import {
 } from 'sentry/utils/discover/charts';
 import getDynamicText from 'sentry/utils/getDynamicText';
 import type {NormalizedTrendsTransaction} from 'sentry/views/performance/trends/types';
-import {getIntervalLine} from 'sentry/views/performance/utils';
+import {getIntervalLine} from 'sentry/views/performance/utils/getIntervalLine';
 
 import {transformEventStatsSmoothed} from '../../../trends/utils';
