+++ static/app/views/performance/trends/chart.tsx
@@ -8,8 +8,7 @@ import {LineChart} from 'sentry/components/charts/lineChart';
 import TransitionChart from 'sentry/components/charts/transitionChart';
 import TransparentLoadingMask from 'sentry/components/charts/transparentLoadingMask';
 import {normalizeDateTimeParams} from 'sentry/components/organizations/pageFilters/parse';
-import type {EventsStatsData, OrganizationSummary, Project} from 'sentry/types';
-import type {Series} from 'sentry/types/echarts';
+import type {OrganizationSummary, Project} from 'sentry/types';
 import {getUtcToLocalDateObject} from 'sentry/utils/dates';
 import {
   axisLabelFormatter,
@@ -21,7 +20,9 @@ import getDynamicText from 'sentry/utils/getDynamicText';
 import {decodeList} from 'sentry/utils/queryString';
 import {useLocation} from 'sentry/utils/useLocation';
 import useRouter from 'sentry/utils/useRouter';
-import {getIntervalLine} from 'sentry/views/performance/utils';
+import generateTrendFunctionAsString from 'sentry/views/performance/trends/utils/generateTrendFunctionAsString';
+import transformEventStats from 'sentry/views/performance/trends/utils/transformEventStats';
+import {getIntervalLine} from 'sentry/views/performance/utils/getIntervalLine';
 
 import type {ViewProps} from '../types';
 
@@ -32,7 +33,6 @@ import type {
   TrendsStats,
 } from './types';
 import {
-  generateTrendFunctionAsString,
   getCurrentTrendFunction,
   getCurrentTrendParameter,
   getUnselectedSeries,
@@ -56,21 +56,6 @@ type Props = ViewProps & {
   transaction?: NormalizedTrendsTransaction;
   trendFunctionField?: TrendFunctionField;
 };
-
-export function transformEventStats(
-  data: EventsStatsData,
-  seriesName?: string
-): Series[] {
-  return [
-    {
-      seriesName: seriesName || 'Current',
-      data: data.map(([timestamp, countsForTimestamp]) => ({
-        name: timestamp * 1000,
-        value: countsForTimestamp.reduce((acc, {count}) => acc + count, 0),
-      })),
-    },
-  ];
-}
 
 function getLegend(trendFunction: string): LegendComponentOption {
   return {
