+++ static/app/views/performance/utils/getPerformanceDuration.tsx
@@ -1,0 +1,5 @@
+import {getDuration} from 'sentry/utils/formatters';
+
+export function getPerformanceDuration(milliseconds: number) {
+  return getDuration(milliseconds / 1000, milliseconds > 1000 ? 2 : 0, true);
+}
